CLI: Document and move `EnsureDaemonReady()` into partial relating to IPC instead
